[C++][CMake] Move frame.hpp from spatial to multibody
